+++ data/authors.json
@@ -1,0 +1,18 @@
+[
+  {
+    "first_name": "Carlos",
+    "last_name": "Vaca"
+  },
+  {
+    "first_name": "Goku",
+    "last_name": "Kasla"
+  },
+  {
+    "first_name": "Hi",
+    "last_name": "Ho"
+  },
+  {
+    "first_name": "Hi",
+    "last_name": "lo"
+  }
+]
