+++ data/labels.json
@@ -1,0 +1,18 @@
+[
+  {
+    "title": "Book",
+    "color": "Blue"
+  },
+  {
+    "title": "Hello",
+    "color": "Black"
+  },
+  {
+    "title": "Lo",
+    "color": "Blue"
+  },
+  {
+    "title": "LL",
+    "color": "White"
+  }
+]
